Refactored app codebase
